requirements.txt
requests==2.33.0
pillow == 12.1.1
